move nacp stuff to seperate file and better main code
